+++ components/hxlight_ble_adv/__init__.py
@@ -43,7 +43,7 @@ CONFIG_SCHEMA = cv.All(
                 cv.positive_time_period_milliseconds,
                 cv.Range(min=TimePeriod(milliseconds=20), max=TimePeriod(milliseconds=10240)),
             ),
-            cv.Optional(CONF_ADV_DURATION, default="1000ms"): cv.All(
+            cv.Optional(CONF_ADV_DURATION, default="800ms"): cv.All(
                 cv.positive_time_period_milliseconds,
                 cv.Range(min=TimePeriod(milliseconds=50), max=TimePeriod(milliseconds=10000)),
             ),